+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 master (unreleased)
 -------------------
 
+- Queryable properties can now be populated in raw queries by using the property name as SQL column name
+
 1.8.5 (2023-11-13)
 ------------------